extracted validation code
